FreeDNS

+++ NS_Install2.sh
@@ -1,42 +1,26 @@
 #!/bin/bash
 
 echo
-echo "Log into noip.com"
+echo "Finalizing Nightscout installation"
 echo
 
-if [ ! -s /srv/nightscout-vps ]
+if [ "$(node -v)" = "" ] # If Node.js is not installed
 then
-cat > /tmp/install2_note << EOF
-Complete Initial Nightscout installation first.
-
-EOF
-cd /tmp
-dialog --textbox install2_note 6 51
+clear
+dialog --colors --msgbox "     \Zr Developed by the xDrip team \Zn\n\n\
+You need to complete installation phase 1 first." 9 50
 exit
 fi
 
-if [ ! -s /usr/local/etc/no-ip2.conf ]
-then
-cd /usr/src
-sudo tar -xzf /srv/nightscout-vps/helper/noip-duc-linux.tar.gz
-cd /usr/src/noip-2.1.9-1
-sudo make install
-else
-echo "Noip client already installed - delete /usr/local/etc/no-ip2.conf if you want to change config"
-fi
-noip2 -S
+Test=0
+#Test=1 ########################## This line must be commented out before submitting a PR. ###################
 
-hostname=`noip2 -S 2>&1 | grep host | tr -s ' ' | tr -d '\t' | cut -f2 -d' ' | head -1`
-
-if [ "$hostname" = "" ]
+if [ "`id -u`" != "0" ]
 then
-echo "Could not determine host name - did no ip dynamic dns install fail?"
-echo "Cannot continue!"
+echo "Script needs root - use sudo bash NS_Install2.sh"
+echo "Cannot continue.."
 exit 5
 fi
-
-# execute installer
-noip2
 
 sudo apt-get install -y nginx python3-certbot-nginx inetutils-ping
 
@@ -54,7 +38,7 @@ echo "Nginx config already patched"
 fi
 
 sudo service nginx start
-sudo certbot --nginx -d "$hostname" --redirect
+# sudo certbot --nginx -d "$hostname" --redirect  # We are doing this in ConfigureFreedns.sh
 
 sudo systemctl daemon-reload
 sudo systemctl start mongodb
@@ -62,6 +46,9 @@ sudo systemctl start mongodb
 echo
 echo "Setting up startup service"
 echo
+
+if [ ! -s /etc/nsconfig ] # Only if nsconfig does not exist already
+then
 
 cat > /etc/nsconfig << EOF
 
@@ -73,47 +60,70 @@ export DEVICESTATUS_ADVANCED="true"
 
 EOF
 
-cat > /etc/nightscout-start.sh << "EOF"
+fi
 
+if [ $Test -lt 1 ] # If we are not testing.
+then
+
+cat > /etc/nightscout-start.sh << "EOF"
 #!/bin/sh
 export PATH="/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
-
 . /etc/nsconfig
-
 export MONGO_COLLECTION="entries"
 export MONGO_CONNECTION="mongodb://username:password@localhost:27017/Nightscout"
 export INSECURE_USE_HTTP=true
 export HOSTNAME="127.0.0.1"
 export PORT="1337"
-
 cd /srv/nightscout-vps
-
 while [ "`netstat -lnt | grep 27017 | grep -v grep`" = "" ]
 do
 echo "Waiting for mongo to start"
 sleep 5
 done
 sleep 5
-
 while [ 1 ]
 do
 node server.js
 sleep 30
 done
-
 EOF
 
-echo
-echo "You can edit the full configuration with: sudo nano /etc/nsconfig"
-echo
+else # We are testing.
+cat > /etc/nightscout-start.sh << "EOF"
+#!/bin/sh
+export PATH="/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
+. /etc/nsconfig
+export MONGO_COLLECTION="entries"
+export MONGO_CONNECTION="mongodb://username:password@localhost:27017/Nightscout"
+export INSECURE_USE_HTTP=true
+export HOSTNAME="127.0.0.1"
+export PORT="1337"
+cd /srv/cgm-remote-monitor
+while [ "`netstat -lnt | grep 27017 | grep -v grep`" = "" ]
+do
+echo "Waiting for mongo to start"
+sleep 5
+done
+sleep 5
+while [ 1 ]
+do
+node server.js
+sleep 30
+done
+EOF
+
+fi
 
 cs=`grep 'API_SECRET=' /etc/nsconfig | head -1 | cut -f2 -d'"'`
 
 echo "Current API secret is: $cs"
-
 echo
 echo "If you would like to change it please enter the new secret now or hit enter to leave the same"
 
+for loop in 1 2 3 4 5 6 7 8 9
+do
+read -t 0.1 dummy
+done
 read -p "New secret 12 character minimum length (blank to skip change) : " ns
 
 if [ "$ns" != "" ]
@@ -133,22 +143,13 @@ fi
 
 cat > /etc/rc.local << "EOF"
 #!/bin/bash
-
 PATH="/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin"
-
 cd /tmp
-
 swapon /var/SWAP
-
 service snapd stop
-
-/usr/local/bin/noip2 &
 service mongodb start
-
 screen -dmS nightscout sudo -u nobody bash /etc/nightscout-start.sh
-
 service nginx start
-
 EOF
 
 chmod a+x /etc/rc.local
@@ -157,26 +158,28 @@ cat > /etc/systemd/system/rc-local.service << "EOF"
 [Unit]
  Description=/etc/rc.local Compatibility
  ConditionPathExists=/etc/rc.local
-
 [Service]
  Type=forking
  ExecStart=/etc/rc.local start
  TimeoutSec=0
  StandardOutput=tty
  RemainAfterExit=yes
-
 [Install]
  WantedBy=multi-user.target
-
 EOF
+
 sudo sed -i -e 'sX//Unattended-Upgrade::Automatic-Reboot "false";XUnattended-Upgrade::Automatic-Reboot "true";Xg' /etc/apt/apt.conf.d/50unattended-upgrades 
 sudo systemctl daemon-reload
 sudo systemctl enable rc-local
 
-echo
-echo "Starting everything up - if works also check okay after a reboot"
-echo
-
 sudo systemctl start rc-local.service
+ 
+sudo /xDrip/scripts/ConfigureFreedns.sh
+if [ ! -s /tmp/FreeDNS_Failed ]
+then
+clear
+dialog --colors --msgbox "     \Zr Developed by the xDrip team \Zn\n\n\
+Press enter to restart the server.  This will result in an expected error message.  Wait 30 seconds before clicking on retry to reconnect or using a browser to access your Nightscout." 10 50
 sudo reboot
+fi
